NET-105: move crypto utils
